add the draw history functionality in the frontend
This commit also removes one of the websocketmessageutils since it made
it really hard to use and hides out some important details

Now when making a message, it is more clear what needs to happen
especially when you have the draw.go struct on another window. It is
clear what the struct in the frontend should look like.